Fixed issue with variadic log macro and some warnings with pedantic flag

+++ lib/io.c
@@ -4,7 +4,7 @@
 #include <io.h>
 
 int ssend(int socket, char *message){
-    int sentBytes = send(socket, message, strlen(message), 0);
+    size_t sentBytes = send(socket, message, strlen(message), 0);
 
     if(sentBytes != strlen(message))
         log(FATAL, "Failed to send message")
